Autofocus cancel button
Makes it easier using the popup with keyboard only. The cancel button is autofocused, so that pressing enter will close the popup. OK button is just one tap of the TAB key away. The reason the OK button is not autofocused is because accidentally pressing enter could then potentially cause harm.