Add option for button autofocus
The focus property of the options object must be "ok" to focus okButton, "cancel" to autofocus the cancelButton, and any other non-empty string for none of the buttons to be focused. The default is "cancel". The _options object is added to this in order for _init to access the options. The README is updated.